Migrate attachments to its own file
