+++ nylas/models/attachments.py
@@ -1,0 +1,24 @@
+from dataclasses import dataclass
+from typing import Optional
+
+from dataclasses_json import dataclass_json
+
+
+@dataclass_json
+@dataclass
+class Attachment:
+    """
+    An attachment on a message.
+
+    Attributes:
+        id: Globally unique object identifier.
+        size: Size of the attachment in bytes.
+        filename: Name of the attachment.
+        content_type: MIME type of the attachment.
+    """
+
+    id: str
+    size: int
+    filename: Optional[str] = None
+    content_type: Optional[str] = None
+
